Implemented a more general approach for player detection
Previously the user had to specify what player they would like to use by
setting the player variable. Now the preferred_player variable should be set,
if the preferred player is running the widget will interact with it. if it
isin't it will interact with whatever playerctl detects is playing something
when passing it %any.